refactor: CCP-4535 move user management containers
